Refactored the menuBar class into correct package
Also bugfixed the logout
