+++ src/abs/controller/OwnerController.java
@@ -13,8 +13,8 @@ import abs.model.Customer;
 import abs.model.Employee;
 import abs.model.Owner;
 import abs.model.User;
-import abs.view.ABSMenuBar;
 import abs.view.AppFrame;
+import abs.view.GUIComponents.ABSMenuBar;
 import abs.view.owner.AddAvEmployee;
 import abs.view.owner.AddEmployee;
 import abs.view.owner.BookingForCustomer;
